kibana marvel: settings unauth handling
When the user is not authorized to access the Settings resources on the
server, the request will be left in pending and never return. When this
happens in a page load, usually there is some other resource that is
unauthorized and will cause a fatal error in the UI.

This just cleans the response so it is not left in pending.

Original commit: elastic/x-pack@69265c0d1a8f54ef8bc2930bfbc90e3c1360cdd0