kibana marvel - remove get_nodes_for_clusters
The query done in this module was nearly identical to the query in get_clusters_health. So I added to get_clusters_health to produce the kind of node data that get_nodes_for_clusters was producing.

The node data is more simplified now though. It doesn’t need to be fully detailed because the UI only uses this array to show a count of nodes per cluster.

Original commit: elastic/x-pack@97d7ae19b1be2f7c47da51685ed730d5b15856b1