monitoring ui: always return cluster_state object
This change is needed because in Elasticsearch 3.0, bad query DSL syntax would be ignored for times when the index was not found. Returning undefined for the last state didn’t matter before, because queries that used it were also using `.kibana-devnull` as the index, and the error reported by Elasticsearch would be an `index_not_found_exception` with a 404 status code.

In Elasticsearch 3.0, exceptions in the query syntax supersede an `index_not_found`. An exception was happening when the last cluster state was `undefined` instead of an object, because the query had syntax such as `{ “term”: {} }`. So anywhere there might be invalid query syntax needs to be fixed, even if the index in the query is invalid, because an invalid index doesn’t hide the invalid syntax query exception.

Original commit: elastic/x-pack@73667fbb9c1aad6712be81d9401451836443648c