monitoring ui: remove field validation relationships
Not practical to enforce these relationships because if we are
defaulting to use the Kibana config for Elasticsearch connection, then
the config properties get copied one at a time. Because of that, we
can't well enforce when multiple properties should always appear
together. Because order matters in how the properties are registered,
and order is not enforceable.

Original commit: elastic/x-pack@3c357732ef8534feae5f2eeacf6e882dcf3b5fba