Kibana/Shield: Do not prefix message in wrapped errors
The default behavior of Boom.wrap is to use the message of an error
object as the message in the resulting payload for 400 level errors, so
when we pass that information along manually in our wrapError utility,
Boom assumes you are trying to create a prefixed error message, so it
ends up duplicating the message.  For example, if your error had a
message of "Invalid", the resulting boom error message would be
"Invalid: Invalid". This behavior is fixed by making our wrapError
utility not pass along a duplicate error message.

In addition to fixing this bug, the wrapError utility was moved into the
errors module.

Original commit: elastic/x-pack@2f4e608307391b604e5c62f6967f16ecef3351b4