Renaming xpackMain plugin folder from 'main' to 'xpackMain'
Currently the xpackMain plugin does not expose any public (client-side) files. If
it ever wanted to do this, it would need to provide a `uiExports.app.main` property.
The value of this property needs to be `plugins/{plugin-id}/...`. In other words, the
`{plugin-id}` in that folder path needs to match the actual `id` of the plugin. Therefore
this change.

Original commit: elastic/x-pack@da8bb62e428df73481d479e9db46b043f2287781