[Monitoring UI] Change "Status" to "Cluster" for Cluster Summary
This is an extremely minor change that changes the word "Status" to "Cluster" in the cluster level summaries for
both Elasticsearch and Kibana. This helps to provide context not only for the Status itself, but the rest of the
line's stats.

For such a small change, I'm very happy with how it changes my initial response to seeing it.

Original commit: elastic/x-pack@8497754c1fe450fd0803bf9f578620ff99b05a5c