[Monitoring UI] Add Overview Tab to Elasticsearch page
This adds the Overview tab that was missing from Elasticsearch cluster page. The page already existed as the default selection, but now it's obvious how to find it without having to click on the "Elasticsearch" breadcrumb to get back to it.

- This also removes kbn-href from any "tab as title" value so that they don't become clickable (wasted self-redirects given that the page is auto-updated).

Original commit: elastic/x-pack@790c51a2330530e8c8c28437d1873aef04cbe92f