[Monitoring UI] Remove "Overview" Label from Cluster Name
This removes the "Overview: " label from prefixing the cluster name for any selected cluster.

The only downside to this change, to me, is that "elasticsearch" will likely be mistaken for a product name rather than a cluster name by newcomers that are just starting out. We could possibly add an icon to distinguish that it's a "thing" like an index or a node is a thing.

Original commit: elastic/x-pack@6335ee92b02315df20329bbaca22f1c32ec2dfcf