[Monitoring UI] Remove Charts from "Indices" Tab
This makes the "Indices" and "Nodes" tabs behave in the exact same way: a list of indices or nodes.

All of the charts on the "Indices" tab are duplicated on the "Overview" tab in every way. By removing them from the indices tab it makes it _much_ easier to navigate to individual indices and it entirely avoids unnecessary duplication.

Note: this matches the latest behavior of https://projects.invisionapp.com/d/main#/console/4925467/164594676/preview .

Original commit: elastic/x-pack@62b2025fb773612abeb7cf00f5a7eb48212cb544