[Monitoring UI] Fix index listing for Not Available Index
If any index happens to exist in the listing results because of the time picker, but its status is "Not Available" (because it's deleted), then it won't have any metric data supplied. Before this, this caused an unhandled error; this allows that scenario.

Original commit: elastic/x-pack@7d9c6c1fecc759d5b64a814fcec540f67d5c1cee