security: Auth redirects determined by kbn-version header
Security must redirect users to the login page when a user is doing
full-page navigation (i.e. refreshing the page or loading Kibana from
scratch), but it needs to reject xhr requests with a 401. Since there is
no definitive way to check for xhr requests, we were relying on there
being "html" somewhere in the accept header to make that determination.

However, the accept header is insufficient as some configurations of IE
do not include "html" anywhere in their accept header, so those users
would just get full screen exceptions. The kbn-version header is used
for all xhr requests for version matching and xsrf protection, so it
should be a much better determination of whether a request should 401 or
not.

Original commit: elastic/x-pack@3e6ae9f02e5ad2cb972b078fd6cf7b00876b3448