Initialize lastCheckedOn in local storage, if necessary
Do this in a module.config function rather than in the module.run function to ensure that the initialization takes place before any UI elements are available to the user. This will ensure that all reports that complete in this browser session (but none before then) are "caught" for notifications.

Original commit: elastic/x-pack@e24a220b974d03325bc683f8083b9a091beff049