[Monitoring UI] Use Node's Persistent ID as node resolver
This adds the UUID as a valid node resolver option now that it is persistent (it used to be ephemeral, thus changed with each node restart) and makes it the default resolver.

With this, we can start to look at simply removing the configurable node resolver feature altogether.

Original commit: elastic/x-pack@dd02dcc2db1218d7ad049e4ab0051107d8a0a39f