[Monitoring UI] Add Points to Charts
This turns on points for all charts within Monitoring.

I also enabled autoHighlight, which highlights the closest point to the mouse (within both X and Y axes). This led down a rabbit hole of improving how _we_ select values to display in the legend. It used to be that we always showed the first Y value whose corresponding X value was less than that of the cursor (x < cursor_x). This changes it so that the first point to be greater than the cursor's position causes it to look at the previous value's distance from that cursor to see which is better. Given equal distances (both are exactly halfway away), then it will choose the one that is after the cursor. It will only do this until it finds a point in _any_ series within a plot rather than for _each_ series in the plot, then it shares the same index.

It also uses the highlighted "item" as the given index to highlight in all legends, which is much faster than looping across every series at all, but this is only possible when the mouse causes a point to be visually highlighted.

Original commit: elastic/x-pack@d15a8fcc813c6e2df6abbaff1d1e8fd836bc3591