security-ui: safely check for api tag in auth redirect
There is no guarantee that the request object will contain the settings
hash with tags, so we perform a safe check for that value to ensure
we're not throwing an exception when it does not exist.

The test fixtures are updated to reflect that uncertainty.

Fixes test failures that were introduced in elastic/x-pack#3172

Original commit: elastic/x-pack@3e8ad947a8bd17a43973147acc59f647c30ca45b