Check new kbn-xsrf header in auth redirect
When determining whether a request is an api request or not, we now also
check the presence of the new kbn-xsrf header as one possible indicator.

This does not change the behaviors for the kbn-version header nor the
api tags concept.

Original commit: elastic/x-pack@f6168855688f5351768630415bc1b0519b75866a