reporting: Require user to save changes for reporting
The `appStatus` is checked to see if the state of the app is "dirty" or
not, and a message is rendered in place of the reporting actions if it
is. This should mean that any time a user has unsaved changes, they must
save changes before they can generate a new report.

Original commit: elastic/x-pack@aeca5f2db99626992680c2d34811fa172903929d