[Monitoring UI] Remove 'meta' parameter
This removes the `meta` parameter specified with monitoring requests to the cluster, which was intended to make identifying requests easier with verbose trace logging. We can reintroduce this if we can remove the parametr from being sent to the cluster.

Original commit: elastic/x-pack@d98666e73a9b04e122e599a067313db544b20175