Reverting change made in elastic/x-pack@9942e00 to this file as it wasn't having the desired effect
The change made in elastic/x-pack@9942e00 was supposed to ensure that the code in the function passed to uiRoutes.addSetupWork ran before the route was resolved (i.e. the code in the Graph app was executed). In testing I found that that code in the Graph app was being run first instead.

Original commit: elastic/x-pack@3af146f6b8a030e58d7cb132079743c9ca43f581