[Monitoring UI] Combine Advanced Node Charts (elastic/x-pack#4144)
This combines the Read thread charts, as well as Index Time and Throttling
time charts. It also adds the Search Total to the Indexing Rate chart,
which is now renamed the Request Rate.

This drops the "Store Throttle", which is no longer sent in 5.0.

Original commit: elastic/x-pack@29816d588709faf62623963d86f11aa8207fe4fb