[Monitoring UI] Default Calculation should ignore negatives (elastic/x-pack#4146)
* [Monitoring UI] Default Calculation should ignore negatives

This ignores negative numbers by default, instead replacing them with zero
values. This makes it so that derivatives are displayed properly when the
value is reset in the middle of the chart (e.g., due to a node restarting,
which resets its counter).

Original commit: elastic/x-pack@a5c08a8557be6566967ae912ab5c9b43003eda25