Wait for visualizations to render, avoid timeout where possible (elastic/x-pack#4105)
* clean up the css injection code

* wait for visualizations to render

first wait for them to exist in the DOM, then listen for the rendered events

* use loadDelay when no evented visualization is found

* listen on visualize element

updated to work with the changes in Kibana PR elastic/x-pack#8914

* fix passing in the load delay

used when visualizations don't fire the renderComplete event

* separate the listener function

easier to read the code when the function is defined separate from ph.evaluate

* add check for implementsRenderComplete

also track every visualization separately, so that 1 vis firing multiple times won't complete the wait

* use the new onRenderComplete method

this fires immediately if the event already happened

* remove redundant check for 0 visualizations

also remove value from the resolve call, it's not used and isn't helpful

* define function before it gets used

don't rely on hoisting, per the styleguide

* make ph.evaluate use consistent

also unwrap the css injection code

* use Promise.all instead of tracking visualization event counts

* wait longer for kibana to load, add waitForSelector

now that the code actually waits for kibana to start rendering on the page, so when running at max concurrency, errors were much more frequent. a waitForTime was added, which waitFor now uses, but it allows the caller to pass in a custom timeout. waitForSelector was also added to streamline code and usage, which also takes the time. phantom now waits up to 30 seconds for kibana to load, regardless of the capture.timeout setting is. waitForSelector is also used in the screenshot code now, which makes the intention clearer

* make waitForTime recursive

by making the function call itself once it's done with evaluate, this removes the issue where there could be multiple instances running at the same time, so it no longer has to track resolution, and should be easier on system load

* handle non-error rejections

also add 'Unspecified error' text to empty rejections

* add a capture.settleTime config

* use render-count attributes instead of events

updated for kibanaelastic/x-pack#9176. also passes the settleTime and uses it to wait for the renderCount value to settle.

* update the docs

* move max load timeout to constants

* update completion code to use render-counter

new directive added in elastic/kibanaelastic/x-pack#9206

* simply wait for application, not visualize

Original commit: elastic/x-pack@36e03339f9768c70e310cdaaec71b907059ab3ab