[Monitoring UI] Debounce Crosshair Movement (elastic/x-pack#4227)
* [Monitoring UI] Debounce Crosshair Movement

This really improves the performance of crosshair (the red line) movement across the charts on the screen and improves efficiency of the UI overall. It also guarantees that the crosshair on other charts will end up at the same position after the next tick of the debounce.

Original commit: elastic/x-pack@097f0e4681ab5855824454e51bb466212836f66b