[Monitoring UI] Fix table data filtering (elastic/x-pack#4299)
The `getFilteredData` function shouldn’t just take the entire row object
and flatten it into a filterable string, because there are a lot of
fields for metadata in the row object. Instead, we add a new option for
table data called `filterFields` that’s used to pick out the row data
and use that for string matching in the filter.

In the future, we can add a tooltip or help text for the tables that
explain which columns and values can be used for filtering. This change
doesn’t add that, but it helps us get a little closer.

Original commit: elastic/x-pack@3dc156ab0bc40c5c8e259e16aadf9ad199f30d07