[Monitoring UI] Only Publish Latest Kibana Stats (elastic/x-pack#4290)
* [Monitoring UI] Only Publish Latest Kibana Stats

The hapi monitoring feature that we use, uses ops.interval to decide when to invoke the monitoring callback. This value was switched from `10s` (10000) to `5s` (5000). As a result, we receive two updates per flush operation that we perform asynchronously on our own timer. This change will rollup any values that need rolled up (summed / maximum taken from the `n` updates).

This change simplifies it to slightly decouple the two timers. First, it removes the queuing that was in place that enabled this in the first place, in favor of a simpler approach with a singly-set variable (`lastOp`). The latest "push" of monitoring data now sets the lastOp, and that's it. No array business or memory.

The downside of this is that we no longer have any concept of a queue, but we used to wipe it out regardless of success or failure anyway, so getting those benefits will require the same amount of work. The upside is that we save a little memory and send less data.

Original commit: elastic/x-pack@ba62126f5bac78aca6943e9efb27ab5dc87a63c9