[Monitoring UI] Disable most routes with monitoring.ui.enabled: false (elastic/x-pack#4820)
This effectively disables the server side functionality based on the user's settings, rather than actually running everything and selectively ignoring it. Conveniently and intentionally, this breaks UI routes for Monitoring pages.

A big exception is that as long as the user leaves on stats reporting,
then we must leave access open to the clusters server route (so that it
can fetch the phone home data). This means that the top-level cluster
pages are still accessible if you manually go to Monitoring.

Original commit: elastic/x-pack@038a48b26363a592982cbe8d101237a86d0a537a