Graph UI enhancement - better “add links to existing terms” functionality. (#434)
The old logic for adding links to existing vertices sometimes failed to discover links that were present in the search index because of the way it worked with samples of data. The new logic used here looks at all the relevant data rather than a sample using the new adjacency_matrix aggregation. Due to to limitations of memory we select a max of 100 vertices at a time to consider for link discovery. If more than 100 vertices are selected in the UI then we pick a random sub-selection of 100 vertices.
Also this change removes some dead code from the “expand” function now that we use the _search API rather than the Graph API to look for links.

Closes #4896