Gradle daemon is a demon
This commit adds a property that will prevent the Gradle daemon from
being used for builds (even if one is running). This is to avoid some
nasty issues (e.g., SIGBUS faults and other mmap diasters) that result
from class loaders not being closed properly, or issues like the changes
in the environment not propagating to commands executed by the daemon.