[ML] Resizing charts based on Y-axis values (#925)
* [ML] Resizing charts based on Y-axis values

Fixes issues where the y-axis tick labels would be partially hidden if they were too long.
Now the highest value is measured and the width of the chart's left margin adjusted to accomodate it.

In the multi-metric page all charts are aligned, so the margins all need to be enough to display the largest number from any of the other charts.

Fixes #752

* [ML] Hide split cards if no fields selected