[ML] Changing Open link menu button (#1236)
The "Open link" dropdown control does not need to be a button element as it looks nothing like a button.
Currently we are giving the button the `kuiButton` class and then overriding the style to make it look like a link.
This change turns the control into a clickable span element and removes the `kuiButton` and `kuiButtonGroup` classes.

Closes #1034