[ML] Allow field names with bad characters (#1267)
Fixes issue where multi and single metric wizards fail if the data contains field names which contain non-alphanumeric characters

In the single metric page, because of the use of aggregations in the datafeed config, any bad field names are replaced by a dummy name in the detectors.
e.g. field_0, field_1

Fixes #1237