[ML] Adding index loaded validation to advanced job (#1296)
When cloning a job, if the user doesn't have access to the index a user friendly warning is displayed when they try to save.

The check looks to see if the types have been loaded correctly.
This check will need to change whe types are removed from ES.
However, that work will be covered in the rest of the changes needed to remove types from the job creation pags.

Fixes #701