[Monitoring] Replace msearch with Field Collapsing in get_cluster_health (#1323)
This replaces the `msearch` (one search per cluster) with a single search that does the exact same thing by using field collapsing.

By making the change, it removes N index level searches in favor of a single one. This means that we are most likely removing `7N` shard searches in favor of `7` shard searches (assuming 1 index + 1 shard per day with 7 days available to search, which is the default)!