[ML] Bucket span estimator handle long poll intervals (#1556)
* [ML] Bucket span estimator handle long poll intervals

Currently if a poll is found and it is greater than 3 hours, all tests are skipped.
This fix adds additional intervals: 6h, 12h, 1d ... 7d
Now if a poll interval is greater than 3h, it will determine which of the long intervals the poll interval is closest to and then return, skipping all other tests.

* [ML] Adding missing 3h interval

* [ML] Fixing typo in comment