[Monitoring] Support Future Timeranges in Charts (#1800)
This enables the time picker to be used to select partially future time
ranges without confusing the charts.

The change is that we no longer create "extended bounds" buckets for the
future (or past) time ranges that did not exist in the data set. Instead,
we filter on the time range that we have and get only the data exists.
From there, we explicitly set the x-axis min/max bounds to the associated
time range at the time of the request (as opposed to what it happens to be
_right now_).

This saves data in the response payload, but it more importantly fixes a
long-standing issue where the chart's legend always showed `N/A` for
data that didn't exist because they had something like "Today" selected
for the time range.