[Phone Home] Allow Kibana Monitoring to start while we detect Clouds (#1937)
This changes the cloud service detection to happen asynchronously. The advantage is that the Kibana monitoring will use the cloud details, if they ever appear, but it will otherwise start immediately.

If we wait to start before sending stats, then we guarantee we send the cloud details with all monitoring docs, but we can wait at least 45s (in my local tests) when not running on a cloud service for requests to time out.