[Upg Asst] Allow canceling another instance's task (#2063)
Conflicts:
      plugins/upgrade/public/lib/reindex/step_methods.js

* [Upg Asst] Resurrect refresh indices button
* [Upg Asst] Clean up completed indices
* [Upg Asst] Simplify loadIndices
* [Upg Asst] Allow canceling another instance's task
* [Upg Asst] Fix: incorrectly adding step state
* Fix a few bugs
* [Upg Asst] Lightly handle a weird edge case
When you attempt to cancel a task, sometimes the task doesn't get
canceled at all. It just keeps going after your successful re-
quest to cancel it. When it happens, we shouldn't clear out the
index state of all the steps because the task may come back as
still running. So we'd still want to see all the previously com-
pleted steps. Instead, still call `resetIndex` but not the code
that resets the index state.

* [Upg Asst] Simplify task cancelation
* [Upg Asst] Load task detail on initial load
* [Upg Asst] Prepend routes with `/` for cloud proxy
* [Upg Asst] Show that cancel request is delivered
* [Upg Asst] Stop polling when canceled
* [Upg Asst] Code cleanup
* [Upg Asst] Check for canceled after getting task details