[ML] Checking for detector value undefined (#2367)
When creating a detector, if you press backspace to delete any of the inputs, the react-select component sets the value to undefined rather than an empty string.
It appears this is ignored by the back end when saving a job and so doesn't cause a problem. But it should still not happen.