[Monitoring][WIP] Fix Numerous Layout Issues with Logstash WebCola Layout (#2267)
There are some bugs in the current separated rank function, some configs that cause weird layouts, and some abuses of webcola in the current code.

The general change here is that the layout is much less cluttered, the ranks are more spread out, and there is less criss-crossing of links between nodes. All of these things are a big improvement on the current layout. Additionally, true/false branches are much better aligned now, with the true side almost always on the left (giving consistent presentation) for the overwhelmingly common case.

This also fixes some hard to reproduce bugs in which nodes would sometimes overlap by being more explicit in the constraints. I would screenshot that but its very hard to repro.

Additionally, these constraints are faster to layout for larger configs.

Additionally, I've added a ton of documentation showing how the layout works, which was a major piece of tech-debt given the complexity of webcola.