[ML] Data Preview Loading Indicator (#2489)
This PR solves the issue described in #1734: Adds a loading indicator to the datafeed preview tab in the jobs list table, as well as one for the data preview tab for the advanced job creation.

In addition this also consolidates existing loading spinner code into an angular directive.

The new directive can be used like:

```
<ml-loading-indicator label="Loading" is-loading="isLoading" />
```

The `label` attribute is optional. `is-loading` should be a flag set by an attribute of the encapsulating controller/directive's scope.

This PR replaces existing raw angular/HTML implementations as well as the `chartLoadingIndicator` component. The charts now use the template provided in `plugins/ml/components/loading_indicator/loading_indicator_wrapper.html`. Also includes tests for the loading-indicator directive.