[ML] Make use of overall_buckets API (#2923)
This PR replaces the aggregation search done from the UI in order to determine the scores in the overall swimlane of the anomaly explorer view with a call to the new overall buckets API.

Note that `top_n` is fixed to 1 for now. There is an optimization that considerably speeds up the overall bucket score calculation when `top_n` is 1. If in the future we expose `top_n` to the user, we should revisit to ensure a long call will not delay display the other parts of the anomaly explorer page.