[ML] Introduces State and Listener factory. (#3263)
This PR isolates code originally developed in #3052.

The Listener factory provides the functionality of the event listener function from existing services. It's refactored in a way so it provides a single object with methods exposed to handle events so it's not necessary anymore to create multiple functions to create the event/listener behaviour.

The State factory offers some more convenience to handle state shared amongst directive including persisting state via AppState. It's loosely inspired by the Flux pattern.

The factories in this PR are changed to reflect the pattern introduces with the FieldsServiceProvider, this reduces direct dependency on angular.

This PR doesn't apply this new functionality to existing code, it adds the factories and accompanying tests. Refactoring existing code to use this will be done in some follow up PRs.