[ML] Refactored anomaly table controls. (#3274)
This PR is a follow-up to #3269 and #3263 and refactors the table controls (Severity/Interval dropdown): The dropdowns become their own directives including a service using the new stateFactory. This way we get rid of direct dependencies in the anomalies table directive and are more easily able to move/restructure the position of the table controls. This PR keeps the exact same control positions and functionality, actual changes will be done in a follow up. Additionally, some redundant LESS code related to the severity icons in the dropdown is also refactored (these affects some other files, e.g. watcher related templates)