[ML] Adds the ability to show/hide the anomaly explorer charts. (#3278)
This adds a checkbox to toggle the charts visibility in Anomaly Explorer. The checkbox is only visible if there were actual charts available (e.g. if a cell is selected in one of the swimlanes). The checkbox will remember its state (enabled/disables) even if it was hidden when no charts are available anyway.