Make time picker enable/disable more granular (#3414)
* Make time picker enable/disable more granular

The time picker API in open-source Kibana is becoming more granular. Instead of being able to enable/disable the entire timepicker as a whole, the new API will allow enabling/disabling either the time range selector and/or the auto-refresh selector parts of the time picker individually. This PR updates code in x-pack-kibana to use the new API.

* Using clearer-named API methods

* Missed a couple of spots

* Setting up mock object correctly