[Monitoring] Add back Logstash version for each node in node listing (#3523)
* Add back Logstash version for each node in node listing

This was removed in a prior commit but [a user](https://discuss.elastic.co/t/logstash-version-missing-in-monitoring/109257) reported missing it. Their use case was to sort by versions to easily see which Logstash nodes needed upgrading. While there is a Cluster Alert for mismatched Logstash versions it is not available in X-Pack Basic. So we're bringing back this information in the Logstash Node listing view.

* Using slightly smaller text for the version column