Default to latest pipeline version if none is specified (#3658)
Currently the Pipeline Viewer requires both the pipeline ID and a pipeline hash (representing the version of the pipeline) in order to render that pipeline version. However, we want the Pipeline Viewer to default to the latest version of a given pipeline. We do this by making the pipeline hash parameter optional.