[ML] Fixes an issue with applying the context domain bounds to the focus domain. (#3813)
When the min/max y values of the focus chart are the same, we automatically applied the domain of the context chart if their min/max was different. However, this didn't consider that both charts could have data aggregated on different levels, resulting in different domains. The effect would be that the focus chart's line was in the wrong position or invisible. The fix in this PR additionally checks if the focus chart value is within the context chart's domain, if not, it falls back to the default behavior of just creating a "padding" domain for the focus chart. This PR also removes unnecessary reassignments of focusYScale.