[Monitoring] Show Resolved Alerts when possible (#4064)
This changes the old behavior, which ignored all resolved alerts from being displayed. Now, resolved alerts are displayed _if_ they were resolved within the selected time range.

This change is necessary to enable the usage of the nodes added / removed / restarted cluster alert because that alert is always immediately resolved (because there's no follow-up action that you can depend on).

As a very large benefit, this also increases the utility of the alerts so that we can create this pattern for other alerts, including the ones that already exist. We can change them so that they create a new record whenever they are triggered (after being resolved) so that the user can see the history -- if they care to look at it.

This is the first step toward taking Monitoring to the "alerts" overview screen where alerts drive the UI rather than arbitrary metrics.