test: use 1 shard for graph fixture
The graph api can produce slightly different results depending on how
the data is sharded. Since we're only interested in testing the
interactions with the graph apis themselves in these Kibana tests rather
than testing the quality of the results under varying conditions, we
should just use 1 shard to ensure the results are consistent.