[ML] Add job validation button to jobs list and wizards. (#4677)
Adds the job validation button to the jobs list and the wizards for single metric, multi metric and population jobs.
For each jobService for the wizards, the function getJobFromConfig is now exposed so the directives can use it to transform formConfig to a job configuration before passing it on to the job validation component.
The job validation button is a React/EUI component. The "Create Job" buttons are still angular/HTML, but the styles have been updated to use EUI to match the job validation button style.
The job validation component has been updated with some tweaks, for example to allow a non-filled and/or disabled button style.