[ML] Deduplicate job validation messages. (#4948)
This PR adds a deduplication step to the job validation messages before the messages with parsed templates get returned. The deduplication checks for exact matches of message config objects. If a future update changes the details of a message being returned (e.g. adding a detector name or index to the above messages to make the information more granular) the messages would no longer be deduplicated.